tests/wpt/tests/html/semantics/document-metadata/interactions-of-styling-and-scripting/style-element-media-not-match-does-not-block-script.html

<!DOCTYPE html>
<title>Style element is not script-blocking when media doesn't match</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/utils.js"></script>
<style media="print">
@import url('stylesheet.py?delay=1');
</style>
<h1>Some text</h1>
<script>
test(() => {
  assert_false(styleExists("h1 { color: purple; }"),
               'stylesheet should still be pending');
  const h1 = document.querySelector('h1');
  assert_equals(getComputedStyle(h1).color, 'rgb(0, 0, 0)');
});
</script>
